// web/src/api/model/resourceBinding.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Cubit Control Plane API
 * Strict REST API for Cubit, an open-source bare-metal PaaS for running Cloudflare Workers and Durable Objects using celld, Traefik, and S3 storage.
 * OpenAPI spec version: 1.0.0
 */
import type { ResourceBindingType } from './resourceBindingType';

export interface ResourceBinding {
  type: ResourceBindingType;
  name: string;
  resourceId?: string;
}